requirements.txt
pyodbc
prometheus-client
pyyaml
pyinstaller
